fix db entry update trashing dbidx; other cleanup

+++ src/include/msgparse.h
@@ -12,7 +12,7 @@
 typedef struct {
   dbidx_t   dbidx;
   int       dispidx;
-  int       uniqueidx;
+  long      uniqueidx;
   int       pauseind;
 } mp_musicqupditem_t;
